refactored packaging to new PackageService
