+++ cli/src/main/java/ca/weblite/jdeploy/gui/tabs/CheerpJSettings.java
@@ -77,11 +77,11 @@ public class CheerpJSettings {
         root.setLayout(new BorderLayout(0, 0));
         root.setBorder(BorderFactory.createTitledBorder(BorderFactory.createEmptyBorder(15, 15, 15, 15), null, TitledBorder.DEFAULT_JUSTIFICATION, TitledBorder.DEFAULT_POSITION, null, null));
         content = new JPanel();
-        content.setLayout(new FormLayout("fill:d:grow,left:4dlu:noGrow,fill:d:grow", "center:d:noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
+        content.setLayout(new com.jgoodies.forms.layout.FormLayout("fill:d:grow,left:4dlu:noGrow,fill:d:grow", "center:d:noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
         root.add(content, BorderLayout.CENTER);
         final JLabel label1 = new JLabel();
         label1.setText("<html><p>This section allows you to deploy your application as a web application using CheerpJ and Github Pages</p></html>");
-        CellConstraints cc = new CellConstraints();
+        com.jgoodies.forms.layout.CellConstraints cc = new com.jgoodies.forms.layout.CellConstraints();
         content.add(label1, cc.xy(1, 1));
         final JLabel label2 = new JLabel();
         label2.setText("<html><p>This is experimental, and will only work if your project uses only libraries supported by CheerpJ.</p></html>");
@@ -90,13 +90,13 @@ public class CheerpJSettings {
         label3.setText("<html><p>Currently Swing and AWT UIs are supported.  There is no support currently for JavaFX.</p></html>");
         content.add(label3, cc.xy(1, 5));
         final JPanel panel1 = new JPanel();
-        panel1.setLayout(new FormLayout("fill:d:grow,left:4dlu:noGrow,fill:d:grow", "center:d:grow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
+        panel1.setLayout(new com.jgoodies.forms.layout.FormLayout("fill:d:grow,left:4dlu:noGrow,fill:d:grow", "center:d:grow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
         content.add(panel1, cc.xy(1, 7));
         final JLabel label4 = new JLabel();
         label4.setText("Github Pages Branch");
         panel1.add(label4, cc.xy(1, 5));
         githubPagesBranch = new JTextField();
-        panel1.add(githubPagesBranch, cc.xy(3, 5, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(githubPagesBranch, cc.xy(3, 5, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         enableCheerpJ = new JCheckBox();
         enableCheerpJ.setText("Enable Web App Generation using CheerpJ");
         panel1.add(enableCheerpJ, cc.xy(3, 3));
@@ -104,17 +104,17 @@ public class CheerpJSettings {
         label5.setText("GitHub Pages Branch Path");
         panel1.add(label5, cc.xy(1, 9));
         githubPagesBranchPath = new JTextField();
-        panel1.add(githubPagesBranchPath, cc.xy(3, 9, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(githubPagesBranchPath, cc.xy(3, 9, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label6 = new JLabel();
         label6.setText("GitHub Pages Tag Path");
         panel1.add(label6, cc.xy(1, 15));
         githubPagesTagPath = new JTextField();
-        panel1.add(githubPagesTagPath, cc.xy(3, 15, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(githubPagesTagPath, cc.xy(3, 15, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label7 = new JLabel();
         label7.setText("GitHub Pages Default Path");
         panel1.add(label7, cc.xy(1, 21));
         githubPagesPath = new JTextField();
-        panel1.add(githubPagesPath, cc.xy(3, 21, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(githubPagesPath, cc.xy(3, 21, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label8 = new JLabel();
         label8.setText("E.g. gh-pages");
         panel1.add(label8, cc.xy(3, 7));
