+++ cli/src/main/java/ca/weblite/jdeploy/gui/tabs/DetailsPanel.java
@@ -120,21 +120,21 @@ public class DetailsPanel {
         root.setPreferredSize(new Dimension(640, 500));
         root.setBorder(BorderFactory.createTitledBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10), null, TitledBorder.DEFAULT_JUSTIFICATION, TitledBorder.DEFAULT_POSITION, null, null));
         panel1 = new JPanel();
-        panel1.setLayout(new FormLayout("fill:max(d;4px):noGrow,left:4dlu:noGrow,fill:d:grow", "center:d:noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
+        panel1.setLayout(new com.jgoodies.forms.layout.FormLayout("fill:max(d;4px):noGrow,left:4dlu:noGrow,fill:d:grow", "center:d:noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
         root.add(panel1, BorderLayout.CENTER);
         final JLabel label1 = new JLabel();
         label1.setText("Name");
-        CellConstraints cc = new CellConstraints();
+        com.jgoodies.forms.layout.CellConstraints cc = new com.jgoodies.forms.layout.CellConstraints();
         panel1.add(label1, cc.xy(1, 1));
         name = new JTextField();
-        panel1.add(name, cc.xy(3, 1, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(name, cc.xy(3, 1, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label2 = new JLabel();
         label2.setText("Version");
         panel1.add(label2, cc.xy(1, 3));
         version = new JTextField();
-        panel1.add(version, cc.xy(3, 3, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(version, cc.xy(3, 3, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         title = new JTextField();
-        panel1.add(title, cc.xy(3, 5, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(title, cc.xy(3, 5, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label3 = new JLabel();
         label3.setText("Title");
         panel1.add(label3, cc.xy(1, 5));
@@ -142,7 +142,7 @@ public class DetailsPanel {
         label4.setText("Author");
         panel1.add(label4, cc.xy(1, 7));
         author = new JTextField();
-        panel1.add(author, cc.xy(3, 7, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(author, cc.xy(3, 7, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label5 = new JLabel();
         label5.setText("Description");
         panel1.add(label5, cc.xy(1, 9));
@@ -150,9 +150,9 @@ public class DetailsPanel {
         label6.setText("License");
         panel1.add(label6, cc.xy(1, 11));
         license = new JTextField();
-        panel1.add(license, cc.xy(3, 11, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel1.add(license, cc.xy(3, 11, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JScrollPane scrollPane1 = new JScrollPane();
-        panel1.add(scrollPane1, cc.xy(3, 9, CellConstraints.FILL, CellConstraints.FILL));
+        panel1.add(scrollPane1, cc.xy(3, 9, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.FILL));
         description = new JTextArea();
         description.setRows(3);
         scrollPane1.setViewportView(description);
@@ -160,10 +160,10 @@ public class DetailsPanel {
         panel2.setLayout(new BorderLayout(0, 0));
         root.add(panel2, BorderLayout.NORTH);
         final JPanel panel3 = new JPanel();
-        panel3.setLayout(new FormLayout("fill:max(d;4px):noGrow,left:4dlu:noGrow,fill:d:grow,left:4dlu:noGrow,fill:max(d;4px):noGrow", "center:d:grow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
+        panel3.setLayout(new com.jgoodies.forms.layout.FormLayout("fill:max(d;4px):noGrow,left:4dlu:noGrow,fill:d:grow,left:4dlu:noGrow,fill:max(d;4px):noGrow", "center:d:grow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow,top:4dlu:noGrow,center:max(d;4px):noGrow"));
         root.add(panel3, BorderLayout.SOUTH);
         jarFile = new JTextField();
-        panel3.add(jarFile, cc.xy(3, 3, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel3.add(jarFile, cc.xy(3, 3, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label7 = new JLabel();
         label7.setText("JAR File");
         panel3.add(label7, cc.xy(1, 3));
@@ -192,7 +192,7 @@ public class DetailsPanel {
         selectJarFile.setText("Select...");
         panel3.add(selectJarFile, cc.xy(5, 3));
         homepage = new JTextField();
-        panel3.add(homepage, cc.xy(3, 11, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel3.add(homepage, cc.xy(3, 11, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         verifyButton = new JButton();
         verifyButton.setText("Verify");
         panel3.add(verifyButton, cc.xy(5, 11));
@@ -200,12 +200,12 @@ public class DetailsPanel {
         label10.setText("Repository");
         panel3.add(label10, cc.xy(1, 13));
         repositoryUrl = new JTextField();
-        panel3.add(repositoryUrl, cc.xy(3, 13, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel3.add(repositoryUrl, cc.xy(3, 13, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JLabel label11 = new JLabel();
         label11.setText("Directory:");
         panel3.add(label11, cc.xy(3, 15));
         repositoryDirectory = new JTextField();
-        panel3.add(repositoryDirectory, cc.xy(3, 17, CellConstraints.FILL, CellConstraints.DEFAULT));
+        panel3.add(repositoryDirectory, cc.xy(3, 17, com.jgoodies.forms.layout.CellConstraints.FILL, com.jgoodies.forms.layout.CellConstraints.DEFAULT));
         final JPanel panel4 = new JPanel();
         panel4.setLayout(new BorderLayout(10, 10));
         root.add(panel4, BorderLayout.WEST);
